+++ bin/light-base/src/network_service/tasks.rs
@@ -18,7 +18,7 @@
 use super::Shared;
 use crate::platform::{Platform, PlatformConnection, PlatformSubstreamDirection};
 
-use alloc::{string::ToString as _, sync::Arc, vec};
+use alloc::{string::ToString as _, sync::Arc, vec, vec::Vec};
 use core::{iter, pin::Pin};
 use futures::{channel::mpsc, prelude::*};
 use smoldot::{libp2p::read_write::ReadWrite, network::service};
@@ -121,17 +121,26 @@ pub(super) async fn connection_task<TPlat: Platform>(
     // is done by the user of the smoldot crate rather than by the smoldot crate itself.
     let mut guarded = shared.guarded.lock().await;
     let (connection_id, socket_and_task) = match socket {
-        PlatformConnection::SingleStream(socket) => {
-            let (id, task) = guarded
-                .network
-                .pending_outcome_ok_single_stream(start_connect.id);
+        PlatformConnection::SingleStreamMultistreamSelectNoiseYamux(socket) => {
+            let (id, task) = guarded.network.pending_outcome_ok_single_stream(
+                start_connect.id,
+                service::SingleStreamHandshakeKind::MultistreamSelectNoiseYamux,
+            );
             (id, either::Left((socket, task)))
         }
-        PlatformConnection::MultiStream(socket) => {
-            let (id, task) = guarded
-                .network
-                .pending_outcome_ok_multi_stream(start_connect.id);
-            (id, either::Right((socket, task)))
+        PlatformConnection::MultiStreamWebRtc {
+            connection,
+            local_tls_certificate_multihash,
+            remote_tls_certificate_multihash,
+        } => {
+            let (id, task) = guarded.network.pending_outcome_ok_multi_stream(
+                start_connect.id,
+                service::MultiStreamHandshakeKind::WebRtc {
+                    local_tls_certificate_multihash,
+                    remote_tls_certificate_multihash,
+                },
+            );
+            (id, either::Right((connection, task)))
         }
     };
     log::debug!(
@@ -363,6 +372,8 @@ async fn multi_stream_connection_task<TPlat: Platform>(
     let mut pending_opening_out_substreams = 0;
     // Newly-open substream that has just been yielded by the connection.
     let mut newly_open_substream = None;
+    // `true` if the remote has force-closed our connection.
+    let mut has_reset = false;
     // List of all currently open substreams. The index (as a `usize`) corresponds to the id
     // of this substream within the `connection_task` state machine.
     let mut open_substreams = slab::Slab::<TPlat::Stream>::with_capacity(16);
@@ -424,12 +435,12 @@ async fn multi_stream_connection_task<TPlat: Platform>(
         }
 
         // Perform a read-write on all substreams that are ready.
-        loop {
-            let substream_id = match connection_task.ready_substreams().next() {
-                Some(s) => *s,
-                None => break,
-            };
-
+        // TODO: what is a ready substream is a bit of a clusterfuck, figure out
+        for substream_id in connection_task
+            .ready_substreams()
+            .copied()
+            .collect::<Vec<_>>()
+        {
             let substream = &mut open_substreams[substream_id];
 
             let mut read_write = ReadWrite {
@@ -567,8 +578,15 @@ async fn multi_stream_connection_task<TPlat: Platform>(
         debug_assert!(newly_open_substream.is_none());
         futures::select! {
             _ = message_from_coordinator => {}
-            substream = TPlat::next_substream(&mut connection).fuse() => {
-                newly_open_substream = substream;
+            substream = if has_reset { either::Right(future::pending()) } else { either::Left(TPlat::next_substream(&mut connection)) }.fuse() => {
+                match substream {
+                    Some(s) => newly_open_substream = Some(s),
+                    None => {
+                        // `None` is returned if the remote has force-closed the connection.
+                        connection_task.reset();
+                        has_reset = true;
+                    }
+                }
             }
             _ = poll_after => {}
             _ = data_ready.fuse() => {}