+++ src/libp2p/connection/established/multi_stream.rs
@@ -178,18 +178,18 @@ where
 
     /// Notifies the state machine that a new substream has been opened.
     ///
-    /// `inbound` indicates whether the substream has been opened by the remote (`true`) or
-    /// locally (`false`).
+    /// `outbound` indicates whether the substream has been opened by the remote (`false`) or
+    /// locally (`true`).
     ///
-    /// If `inbound` is `false`, then the value returned by
+    /// If `outbound` is `true`, then the value returned by
     /// [`MultiStream::desired_outbound_substreams`] will decrease by one.
     ///
     /// # Panic
     ///
     /// Panics if there already exists a substream with an identical identifier.
     ///
-    pub fn add_substream(&mut self, id: TSubId, inbound: bool) {
-        let (substream, out_substream_id) = if inbound {
+    pub fn add_substream(&mut self, id: TSubId, outbound: bool) {
+        let (substream, out_substream_id) = if !outbound {
             let out_substream_id = self.next_out_substream_id;
             self.next_out_substream_id += 1;