+++ src/libp2p/connection/noise.rs
@@ -197,6 +197,25 @@ impl UnsignedNoiseKey {
     }
 }
 
+/// Configuration for a Noise handshake.
+pub struct Config<'a> {
+    /// Key to use during the handshake.
+    pub key: &'a NoiseKey,
+
+    /// `true` if this side of the handshake has initiated the connection or substream onto which
+    /// the handshake is performed.
+    pub is_initiator: bool,
+
+    /// Prologue data. The prologue data must be identical on both sides of the handshake,
+    /// otherwise it will fail.
+    ///
+    /// See <https://noiseprotocol.org/noise.html#prologue>.
+    ///
+    /// > **Note**: If a certain protocol specification doesn't mention any prologue, it probably
+    /// >           means that this prologue is empty.
+    pub prologue: &'a [u8],
+}
+
 /// State of the noise encryption/decryption cipher.
 pub struct Noise {
     inner: snow::TransportState,
@@ -506,17 +525,19 @@ enum RxPayload {
 impl NoiseHandshake {
     /// Shortcut function that calls [`HandshakeInProgress::new`] and wraps it into a
     /// [`NoiseHandshake`].
-    pub fn new(key: &NoiseKey, is_initiator: bool) -> Self {
-        NoiseHandshake::InProgress(HandshakeInProgress::new(key, is_initiator))
+    pub fn new(config: Config) -> Self {
+        NoiseHandshake::InProgress(HandshakeInProgress::new(config))
     }
 }
 
 impl HandshakeInProgress {
     /// Initializes a new noise handshake state machine.
-    pub fn new(key: &NoiseKey, is_initiator: bool) -> Self {
+    pub fn new(config: Config) -> Self {
         let inner = {
-            let builder = snow::Builder::new(noise_params()).local_private_key(&key.key.private);
-            if is_initiator {
+            let builder = snow::Builder::new(noise_params())
+                .local_private_key(&config.key.key.private)
+                .prologue(&config.prologue);
+            if config.is_initiator {
                 builder.build_initiator()
             } else {
                 builder.build_responder()
@@ -525,12 +546,12 @@ impl HandshakeInProgress {
         };
 
         // Configure according to the XX handshake.
-        let (tx_payload, rx_payload, rx_messages_remain) = if is_initiator {
-            let tx = Some((1, key.handshake_message.clone().into_boxed_slice()));
+        let (tx_payload, rx_payload, rx_messages_remain) = if config.is_initiator {
+            let tx = Some((1, config.key.handshake_message.clone().into_boxed_slice()));
             let rx = RxPayload::NthMessage(0);
             (tx, rx, 1)
         } else {
-            let tx = Some((0, key.handshake_message.clone().into_boxed_slice()));
+            let tx = Some((0, config.key.handshake_message.clone().into_boxed_slice()));
             let rx = RxPayload::NthMessage(1);
             (tx, rx, 2)
         };
@@ -624,12 +645,6 @@ impl HandshakeInProgress {
             },
             remote_peer_id,
         }
-    }
-
-    /// Returns `true` if the Noise handshake is waiting to write out data. Returns `false` if
-    /// instead it is blocked on incoming data.
-    pub fn ready_to_write(&self) -> bool {
-        !self.tx_buffer_encrypted.is_empty()
     }
 
     /// Feeds data coming from a socket and outputs data to write to the socket.
@@ -843,7 +858,7 @@ pub struct PayloadDecodeError;
 
 #[cfg(test)]
 mod tests {
-    use super::{NoiseHandshake, NoiseKey, ReadWrite};
+    use super::{Config, NoiseHandshake, NoiseKey, ReadWrite};
 
     #[test]
     fn handshake_basic_works() {
@@ -851,8 +866,16 @@ mod tests {
             let key1 = NoiseKey::new(&rand::random());
             let key2 = NoiseKey::new(&rand::random());
 
-            let mut handshake1 = NoiseHandshake::new(&key1, true);
-            let mut handshake2 = NoiseHandshake::new(&key2, false);
+            let mut handshake1 = NoiseHandshake::new(Config {
+                key: &key1,
+                is_initiator: true,
+                prologue: &[],
+            });
+            let mut handshake2 = NoiseHandshake::new(Config {
+                key: &key2,
+                is_initiator: false,
+                prologue: &[],
+            });
 
             let mut buf_1_to_2 = Vec::new();
             let mut buf_2_to_1 = Vec::new();